+++ src/tests/ui/matchers.sh
@@ -12,7 +12,7 @@ collector:
 columns:
   - matchers: [ { type: directory } ]
   - matchers: [ { type: fifo }, { type: socket }, { type: symlink } ]
-  - matchers: [ all: [ type: file, type: executable, not: { glob: "*c" }  ] ]
+  - matchers: [ all: [ type: file, type: executable, not: { glob: c }  ] ]
   - matchers: [ { glob: [ "*.a", "y*z" ] }, { regex: '\A[A-Z0-9]+\z' } ]
   - matchers:
     - all: